+++ lib/surveyor/parser.rb
@@ -1,4 +1,7 @@
 %w(survey survey_translation survey_section question_group question dependency dependency_condition answer validation validation_condition).each {|model| require model }
+
+require 'yaml'
+
 module Surveyor
   class ParserError < StandardError; end
   class Parser
@@ -8,6 +11,9 @@ module Surveyor
     attr_accessor :context
 
     # Class methods
+    def self.parse_file(filename, options={})
+      self.parse(File.read(filename),{:filename => filename}.merge(options))
+    end
     def self.parse(str, options={})
       self.ensure_attrs
       self.options = options
@@ -82,6 +88,7 @@ module Surveyor
           resolve_dependency_condition_references
           resolve_question_correct_answers
           report_lost_and_duplicate_references
+          report_missing_default_locale
           Surveyor::Parser.rake_trace("", -2)
           if context[:survey].save
             Surveyor::Parser.rake_trace "Survey saved."
@@ -115,6 +122,11 @@ module Surveyor
     end
     def block_models
       %w(survey survey_section question_group)
+    end
+    def report_missing_default_locale
+      if !self.context[:survey].translations.empty? && self.context[:survey].translations.select{|t|YAML::load(t.translation)=={}}.empty?
+        Surveyor::Parser.raise_error("No default locale specified for translations.",true)
+      end
     end
     def report_lost_and_duplicate_references
       Surveyor::Parser.raise_error("Bad references: #{self.context[:bad_references].join("; ")}", true) unless self.context[:bad_references].empty?
@@ -177,9 +189,18 @@ end
 # SurveySection model
 module SurveyorParserSurveyTranslationMethods
   def parse_and_build(context, args, original_method, reference_identifier)
+    dir = Surveyor::Parser.options[:filename].nil? ? Dir.pwd : File.dirname(Surveyor::Parser.options[:filename])
     # build, no change in context
     args[0].each do |k,v|
-      context[:survey].translations << self.class.new(:locale => k.to_s, :translation => File.read(Rails.root.join("surveys", v)))
+      case v
+      when Hash
+        trans = YAML::dump(v)
+      when String
+        trans = File.read(File.join(dir,v))
+      when :default
+        trans = YAML::dump({})
+      end
+      context[:survey].translations << self.class.new(:locale => k.to_s, :translation => trans)
     end
   end
 end